+++ raml-model/src/main/java/io/vrap/rmf/raml/validation/InstanceValidator.java
@@ -163,7 +163,7 @@ public class InstanceValidator implements DiagnosticsCreator {
         @Override
         public List<Diagnostic> caseIntegerInstance(final IntegerInstance integerInstance) {
             final List<Diagnostic> validationResults = new ArrayList<>();
-            final Integer value = integerInstance.getValue();
+            final Long value = integerInstance.getValue();
 
             if (typeIs(UNION_TYPE)) {
                 return validationResults;
@@ -179,21 +179,21 @@ public class InstanceValidator implements DiagnosticsCreator {
             }
             if (typeInstanceOf(IntegerTypeFacet.class)) {
                 final IntegerTypeFacet integerType = (IntegerTypeFacet) types.peek();
-                if (integerType.getMinimum() != null && value.compareTo(integerType.getMinimum()) < 0) {
+                if (integerType.getMinimum() != null && value.compareTo(integerType.getMinimum().longValue()) < 0) {
                     validationResults.add(error(integerInstance,"Value {0} < minimum {1}",
                             value, integerType.getMinimum()));
                 }
-                if (integerType.getMaximum() != null && value.compareTo(integerType.getMaximum()) > 0) {
+                if (integerType.getMaximum() != null && value.compareTo(integerType.getMaximum().longValue()) > 0) {
                     validationResults.add(error(integerInstance,"Value {0} > maximum {1}",
                             value, integerType.getMaximum()));
                 }
             } else if (typeInstanceOf(NumberTypeFacet.class)) {
                 final NumberTypeFacet numberType = (NumberTypeFacet) types.peek();
-                if (numberType.getMinimum() != null && value.compareTo(numberType.getMinimum().intValue()) < 0) {
+                if (numberType.getMinimum() != null && value.compareTo(numberType.getMinimum().longValue()) < 0) {
                     validationResults.add(error(integerInstance,"Value {0} < minimum {1}",
                             value, numberType.getMinimum()));
                 }
-                if (numberType.getMaximum() != null && value.compareTo(numberType.getMaximum().intValue()) > 0) {
+                if (numberType.getMaximum() != null && value.compareTo(numberType.getMaximum().longValue()) > 0) {
                     validationResults.add(error(integerInstance,"Value {0} > maximum {1}",
                             value, numberType.getMaximum()));
                 }
